Add a Glow device abstraction
This implements GLOW (GL on Whatever) implementation of the device
abstraction. It is a very close copy of the GL implementation.

This abstraction can be used in multiple environments (Open GL, OpenGL
ES, and WebGL) and avoid target-specific code. This can be used with
wasm32-unknown-unknown + web-sys, stdweb, and natively. I wrote this
to enable limited WebGL 2 + wasm32-unknown-unknown support.

This also makes minor changes to shaders so that they work in WebGL 2
and implements an example that works both in web browsers and natively,
via SDL.